requirements.txt


djoser==2.2.2
dj-database-url==2.3.0
psycopg2-binary==2.9.10
gunicorn==23.0.0
whitenoise==6.7.0
sqlparse==0.5.3

asgiref==3.8.1
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.16.0

packaging==24.2


typing_extensions==4.13.0
tzdata==2025.2